Adds hash-assoc
hash-assoc has the same signature as hash-ref. The difference is
that it returns the key/value pair held by the hash. In particular,
the key that it returns is eq?-equivalent to the one stored in
the hash.

Mutable and immutable hashes have different behavior concerning
which key they retain. A mutable hash always retains the first
key that was used to update the hash. Subsequent updates with
keys that are in the equivalance class of that initial key, as
determined by the equality predicate of the hash, do not affect
which key is stored.

On the other hand, immutable hashes always contain the most recent
key that was used to update the hash.

This patch only adds hash-assoc to racket/c. A similar update is
needed for racket/cs.